+++ flutter/lib/medea_jason.dart
@@ -1,4 +1,6 @@
 export 'src/interface/connection_handle.dart';
+export 'src/interface/enums.dart'
+    show MemberConnectionState, MemberConnectionState_P2P, PeerConnectionState;
 export 'src/interface/exceptions.dart';
 export 'src/interface/media_device_details.dart';
 export 'src/interface/media_manager.dart';