+++ flutter/lib/src/web/connection_handle.dart
@@ -2,12 +2,17 @@
 
 import 'dart:js_interop';
 
+import 'package:js_interop_utils/js_interop_utils.dart';
+
 import '../interface/connection_handle.dart';
 import '../interface/media_track.dart';
 import '../util/move_semantic.dart';
 import 'exceptions.dart';
 import 'jason_wasm.dart' as wasm;
 import 'remote_media_track.dart';
+
+import '../interface/enums.dart'
+    show MemberConnectionState, MemberConnectionStateKind, PeerConnectionState;
 
 class WebConnectionHandle implements ConnectionHandle {
   late wasm.ConnectionHandle obj;
@@ -17,6 +22,18 @@ class WebConnectionHandle implements ConnectionHandle {
   @override
   String getRemoteMemberId() {
     return fallibleFunction(() => obj.get_remote_member_id());
+  }
+
+  @override
+  MemberConnectionState? getState() {
+    return freezeState(obj.get_state());
+  }
+
+  @override
+  void onStateChange(void Function(MemberConnectionState) f) {
+    void fn(JSAny state) =>
+        f(freezeState(state as wasm.MemberConnectionState)!);
+    fallibleFunction(() => obj.on_state_change(fn.toJS));
   }
 
   @override
@@ -60,5 +77,23 @@ class WebConnectionHandle implements ConnectionHandle {
   @override
   void free() {
     obj.free();
+  }
+
+  MemberConnectionState? freezeState(wasm.MemberConnectionState? state) {
+    if (state == null) {
+      return null;
+    }
+
+    if (MemberConnectionStateKind.values[state.kind().toInt()] ==
+        MemberConnectionStateKind.p2p) {
+      return MemberConnectionState.p2p(
+        PeerConnectionState.values[(state.value() as JSNumber).toDartInt],
+      );
+    }
+
+    // TODO: implement for SFU.
+    throw UnimplementedError(
+      'Only MemberConnectionStateKind.p2p is supported.',
+    );
   }
 }